refactor: example page layouts
