+++ lib/examples/prototypes/server-pages/AnError.stories.tsx
