+++ lib/examples/prototypes/server-pages/ProxyError.stories.tsx
